+++ storage/src/postgres/posts.rs
@@ -40,7 +40,7 @@ impl PostDialect for Postgres {
         let existing = sqlx::query_as::<_, (UserId, Option<DateTime<Utc>>)>(
             "SELECT user_id, deleted_at FROM posts WHERE post_id = $1 FOR UPDATE",
         )
-        .bind(i64::from(post_id))
+        .bind(post_id)
         .fetch_optional(&mut *tx)
         .await?;
 
@@ -62,7 +62,7 @@ impl PostDialect for Postgres {
              FROM posts WHERE post_id = $2",
         )
         .bind(now)
-        .bind(i64::from(post_id))
+        .bind(post_id)
         .execute(&mut *tx)
         .await?;
 
@@ -107,7 +107,7 @@ impl PostDialect for Postgres {
         // edit/clear — the column was previously omitted from the SET clause, so
         // an edited summary was silently dropped (surfaced by #545's clear e2e).
         .bind(input.summary.as_ref())
-        .bind(i64::from(post_id))
+        .bind(post_id)
         .fetch_one(&mut *tx)
         .await?;
 
@@ -130,7 +130,7 @@ impl PostDialect for Postgres {
 
         let post_exists: bool =
             sqlx::query_scalar("SELECT COUNT(*) > 0 FROM posts WHERE post_id = $1")
-                .bind(i64::from(post_id))
+                .bind(post_id)
                 .fetch_one(&mut *tx)
                 .await?;
 
@@ -152,7 +152,7 @@ impl PostDialect for Postgres {
 
         let result =
             sqlx::query("INSERT INTO post_tags (post_id, tag_id, tag_display) VALUES ($1, $2, $3)")
-                .bind(i64::from(post_id))
+                .bind(post_id)
                 .bind(tag_id)
                 .bind(tag)
                 .execute(&mut *tx)
@@ -183,7 +183,7 @@ impl PostDialect for Postgres {
             "DELETE FROM post_tags
              WHERE post_id = $1 AND tag_id = (SELECT tag_id FROM tags WHERE tag_slug = $2)",
         )
-        .bind(i64::from(post_id))
+        .bind(post_id)
         .bind(tag_slug)
         .execute(pool)
         .await?
