+++ storage/src/sqlite/media.rs
@@ -14,7 +14,7 @@ impl MediaDialect for Sqlite {
         let row = sqlx::query_as::<_, (ByteSize,)>(
             "SELECT COALESCE(SUM(size_bytes), 0) FROM media WHERE user_id = $1 AND source = 'upload'",
         )
-        .bind(i64::from(user_id))
+        .bind(user_id)
         .fetch_one(pool)
         .await?;
 
@@ -31,7 +31,7 @@ impl MediaDialect for Sqlite {
         let result = sqlx::query(
             "DELETE FROM media WHERE user_id = $1 AND sha256 = $2 AND filename = $3 AND source = $4",
         )
-        .bind(i64::from(user_id))
+        .bind(user_id)
         .bind(sha256)
         .bind(filename)
         .bind(*source)
