+++ storage/src/sqlite/posts.rs
@@ -42,7 +42,7 @@ impl PostDialect for Sqlite {
             let existing = sqlx::query_as::<_, (UserId, Option<DateTime<Utc>>)>(
                 "SELECT user_id, deleted_at FROM posts WHERE post_id = $1",
             )
-            .bind(i64::from(post_id))
+            .bind(post_id)
             .fetch_optional(&mut *conn)
             .await?;
 
@@ -60,7 +60,7 @@ impl PostDialect for Sqlite {
                  FROM posts WHERE post_id = $2",
             )
             .bind(now)
-            .bind(i64::from(post_id))
+            .bind(post_id)
             .execute(&mut *conn)
             .await?;
 
@@ -105,7 +105,7 @@ impl PostDialect for Sqlite {
             // edit/clear — the column was previously omitted from the SET clause, so
             // an edited summary was silently dropped (surfaced by #545's clear e2e).
             .bind(input.summary.as_ref())
-            .bind(i64::from(post_id))
+            .bind(post_id)
             .fetch_one(&mut *conn)
             .await?;
 
@@ -147,7 +147,7 @@ impl PostDialect for Sqlite {
         let result: Result<(), TaggingError> = async {
             let post_exists: bool =
                 sqlx::query_scalar("SELECT COUNT(*) > 0 FROM posts WHERE post_id = $1")
-                    .bind(i64::from(post_id))
+                    .bind(post_id)
                     .fetch_one(&mut *conn)
                     .await?;
 
@@ -169,7 +169,7 @@ impl PostDialect for Sqlite {
             match sqlx::query(
                 "INSERT INTO post_tags (post_id, tag_id, tag_display) VALUES ($1, $2, $3)",
             )
-            .bind(i64::from(post_id))
+            .bind(post_id)
             .bind(tag_id)
             .bind(tag)
             .execute(&mut *conn)
@@ -205,7 +205,7 @@ impl PostDialect for Sqlite {
             "DELETE FROM post_tags
              WHERE post_id = $1 AND tag_id = (SELECT tag_id FROM tags WHERE tag_slug = $2)",
         )
-        .bind(i64::from(post_id))
+        .bind(post_id)
         .bind(tag_slug)
         .execute(pool)
         .await?
